requirements.txt
discord.py
mcrcon
python-dotenv
